+++ src/Model/AcceptedService.java
@@ -20,6 +20,8 @@ import javax.swing.table.DefaultTableModel;
  */
 public class AcceptedService implements Serializable{
     private static final long serialVersionUID = -5285470270309768353L;
+    Listas listas = Listas.getInstance();
+    
     //Servicio
     private Date hora;
     private String Ubicacion;
@@ -127,39 +129,7 @@ public class AcceptedService implements Serializable{
         return ClientPhoneNumber;
     }
     
-    public static void AlarmaAutomatica(){
-        Listas listas = new Listas();
-        String estado = "Aceptado y en Progreso";
-        if(listas.getServiciosAceptados().size()!=0){
-            for (int i = 0; i < listas.getServiciosAceptados().size(); i++) {
-                // Código dentro del bucle
-                long TiempoTranscurrido = tiempoTranscurrido(i);
-                if(TiempoTranscurrido>60&& listas.getServiciosAceptados().get(i).getEstadoDelServicio().equals(estado)){
-                    
-                }else{ //Si Supera el tiempo limite entonces se cancela el servicio:
-                    JOptionPane.showMessageDialog(null,"El servicio de transporte de " +listas.getServiciosAceptados().get(i).getUbicacion()+ "hasta "+listas.getServiciosAceptados().get(i).getDestino() +"se ha cancelado automaticamente debido ha que ha superado el tiempo limite de espera","Advertencia",JOptionPane.WARNING_MESSAGE);
-                    listas.deleteAcceptedService(i);
-                    DefaultTableModel model = new DefaultTableModel();
-                    model.setRowCount(0);
-                    ClientWindow.jTableProgresoServicioUsuario.setModel(model);
-                    DriverWindow.jTableEstadoDelServicioAceptado.setModel(model);
-                    ClientWindow.jLabeEstadoDelPedido.setText("Se ha llegado al limite de espera, pida servicio en otro momento");
-                    ArrayList<AcceptedService> acceptedServices = listas.getServiciosAceptados();
-                    AcceptedServiceTableModel actm = new AcceptedServiceTableModel(acceptedServices);
-                    AdminWindow.jTableFinanzas.setModel(actm);
-                    
-                    
-
-                    //Aqui se va a añadir algo para reportar al conductor por no llegar a tiempo
-                }
-            }
-        }    
-    }
     
-    public static long tiempoTranscurrido(int indice) {
-        Listas listas = new Listas();
-        Date ahora = new Date();      
-        long tiempo = ahora.getTime() - listas.getServiciosAceptados().get(indice).getHora().getTime();
-        return tiempo / 1000; // devuelve el tiempo en segundos
-    }    
+    
+    
 }
